Ensure SpringClientFactory gets used for Feign clients
If the ClientFactory from netflix is used directly it bypasses
all the goodness in SpringClientFactory, so clients created
in Feign before this change were getting the default native
configuration instead of that provided by
RibbonClientConfiguration.

There are no remaining references to the raw ClientFactory in
Spring Cloud, so we probably have seen the last of this.

Fixes gh-118