Jackson2JsonMessageConverter added to Turbine amqp auto-configuration
Message converter is not marked as a @Bean, since the only use is within the configuration class itself. This corresponds with the usage in org.springframework.cloud.bus.amqp.AmqpBusAutoConfiguration.java class.
No container provided ObjectMapper is used for the message converter, as this a JSON to java.lang.String converter.

Fixes gh-126