Default status page and health URLs if possible and where needed
If the management.port is explicitly set to something other than the
server.port, we can use it to set the status page and health URL
default values. (They can still be overridden by config properties.)

If management.port=0 there's not much we can do because we don't know
the port number yet when the instance config is created.

Fixes gh-425