Close inputStream on errors writing to the outputStream
Currently all errors during writing to the outputStream are ignored. So the
inputStream is read to the end. This is a problem when the inputStream is
infite or very long running (`hystrix.stream` for example) and results in the
inputStream never beingClosed and eating up threads (on zuul and the proxied
backend). With this commit the errors during writing are not ignored resulting
in closing the inputStream.