Compose Simple- and DiscoveryClientRouteLocator
With this commit an AbstractZuulRouteLocator is added. This allows us
to remove the logic for simple routes from DiscoveryClientRouteLocator
by not extending the SimpleRouteLocator. The behaviour for the simple
and discovery routes is from now on composed and not inherited. This
follows the "favor composition over inheritance" principle and makes
the code (especially for the DiscoveryClientRouteLocator) a bit less
twisted.