Update RibbonLoadBalancedRetryPolicy.java
//reset same server since we are moving to a new server
sameServerCount = 0;

This reset will make retry to next server equals to retry on same server.

Let's have a example:

RetrySameServer:3
RetryNextServer:1

And we are done with retry on same server.

then condition if(sameServerCount >= lbContext.getRetryHandler().getMaxRetriesOnSameServer() && canRetry(context)) will set to true and sameServerCount = 0; will set to true.

And it will keep retying until sameServerCount does not match. Or If I set to same MaxRetryOnNextServer to 2 it will retry on next server 7 times.